+++ logservice/logpuller/region_request_worker.go
@@ -15,7 +15,6 @@ package logpuller
 
 import (
 	"context"
-	"sync"
 	"sync/atomic"
 	"time"
 
@@ -28,45 +27,6 @@ import (
 // To generate a workerID in `newRegionRequestWorker`.
 var workerIDGen atomic.Uint64
 
-// requestedStore groups region request workers that share one TiKV address.
-type requestedStore struct {
-	storeAddr string
-	// Use to select a worker to send request.
-	nextWorker atomic.Uint32
-
-	requestWorkers struct {
-		sync.RWMutex
-		s []*regionRequestWorker
-	}
-}
-
-func newRequestedStore(storeAddr string) *requestedStore {
-	return &requestedStore{storeAddr: storeAddr}
-}
-
-func (rs *requestedStore) addWorker(worker *regionRequestWorker) {
-	rs.requestWorkers.Lock()
-	defer rs.requestWorkers.Unlock()
-	rs.requestWorkers.s = append(rs.requestWorkers.s, worker)
-}
-
-func (rs *requestedStore) getRequestWorker() *regionRequestWorker {
-	rs.requestWorkers.RLock()
-	defer rs.requestWorkers.RUnlock()
-
-	index := rs.nextWorker.Add(1) % uint32(len(rs.requestWorkers.s))
-	return rs.requestWorkers.s[index]
-}
-
-func (rs *requestedStore) snapshotWorkers() []*regionRequestWorker {
-	rs.requestWorkers.RLock()
-	defer rs.requestWorkers.RUnlock()
-
-	workers := make([]*regionRequestWorker, len(rs.requestWorkers.s))
-	copy(workers, rs.requestWorkers.s)
-	return workers
-}
-
 // regionRequestWorker is responsible for sending region requests to a specific TiKV store.
 // It owns the long-lived request cache and reconnection loop.
 type regionRequestWorker struct {
@@ -74,6 +34,7 @@ type regionRequestWorker struct {
 
 	pd                         pd.Client
 	clusterID                  uint64
+	credential                 *security.Credential
 	pushRegionEvent            func(SubscriptionID, regionEvent)
 	runtimeRegistry            *regionRuntimeRegistry
 	submitDirectFailure        func(regionFailureInfo)
@@ -105,13 +66,12 @@ func (s *regionRequestWorker) handleSessionFailure(session *regionRequestWorkerS
 
 func (s *regionRequestWorker) runNextSession(
 	ctx context.Context,
-	credential *security.Credential,
 ) (*regionRequestWorkerSession, workerSessionFailure, bool, error) {
 	session := newRegionRequestWorkerSession(
 		s.workerID,
 		s.store.storeAddr,
 		s.pd,
-		credential,
+		s.credential,
 		s.clusterID,
 		s.requestCache,
 		s.runtimeRegistry,
@@ -130,10 +90,9 @@ func (s *regionRequestWorker) runNextSession(
 
 func (s *regionRequestWorker) runSessionLoop(
 	ctx context.Context,
-	credential *security.Credential,
 ) error {
 	for {
-		session, sessionFailure, canceled, err := s.runNextSession(ctx, credential)
+		session, sessionFailure, canceled, err := s.runNextSession(ctx)
 		if err != nil {
 			return err
 		}
@@ -160,6 +119,7 @@ func newRegionRequestWorker(
 		workerID:                   workerIDGen.Add(1),
 		pd:                         client.pd,
 		clusterID:                  client.clusterID,
+		credential:                 credential,
 		pushRegionEvent:            client.pushRegionEventToDS,
 		runtimeRegistry:            client.regionRuntimeRegistry,
 		submitDirectFailure:        client.submitDirectFailure,
@@ -169,7 +129,7 @@ func newRegionRequestWorker(
 	}
 
 	g.Go(func() error {
-		return worker.runSessionLoop(ctx, credential)
+		return worker.runSessionLoop(ctx)
 	})
 
 	return worker
