+++ logservice/logpuller/subscription_client.go
@@ -19,23 +19,17 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/pingcap/errors"
-	"github.com/pingcap/kvproto/pkg/metapb"
 	"github.com/pingcap/log"
 	"github.com/pingcap/ticdc/heartbeatpb"
 	"github.com/pingcap/ticdc/logservice/logpuller/regionlock"
 	"github.com/pingcap/ticdc/logservice/txnutil"
 	"github.com/pingcap/ticdc/pkg/common"
 	appcontext "github.com/pingcap/ticdc/pkg/common/context"
-	"github.com/pingcap/ticdc/pkg/config"
-	cerror "github.com/pingcap/ticdc/pkg/errors"
 	"github.com/pingcap/ticdc/pkg/metrics"
 	"github.com/pingcap/ticdc/pkg/pdutil"
 	"github.com/pingcap/ticdc/pkg/security"
 	"github.com/pingcap/ticdc/pkg/spanz"
-	"github.com/pingcap/ticdc/pkg/util"
 	"github.com/pingcap/ticdc/utils/dynstream"
-	kvclientv2 "github.com/tikv/client-go/v2/kv"
 	"github.com/tikv/client-go/v2/oracle"
 	"github.com/tikv/client-go/v2/tikv"
 	pd "github.com/tikv/pd/client"
@@ -191,7 +185,10 @@ type subscriptionClient struct {
 	pdClock      pdutil.Clock
 	lockResolver txnutil.LockResolver
 
-	stores sync.Map
+	requestedStores   *requestedStoreSet
+	regionScheduler   *regionRequestScheduler
+	subscribedSpans   *subscribedSpanSet
+	staleLockResolver *staleLockResolver
 
 	ds dynstream.DynamicStream[int, SubscriptionID, regionEvent, *subscribedSpan, *regionEventHandler]
 	// the following three fields are used to manage feedback from ds and notify other goroutines
@@ -201,23 +198,21 @@ type subscriptionClient struct {
 
 	// the credential to connect tikv
 	credential *security.Credential
+}
 
-	totalSpans struct {
-		sync.RWMutex
-		spanMap map[SubscriptionID]*subscribedSpan
+func (s *subscriptionClient) ensureHelpers() {
+	if s.requestedStores == nil {
+		s.requestedStores = newRequestedStoreSet(s)
 	}
-
-	// rangeTaskCh is used to receive range tasks.
-	// The tasks will be handled in `handleRangeTasks` goroutine.
-	rangeTaskCh chan rangeTask
-	// regionTaskQueue is used to receive region tasks with priority.
-	// The tasks will be handled in `handleRegions` goroutine.
-	regionTaskQueue *PriorityQueue
-	// resolveLockTaskCh is used to receive resolve lock tasks.
-	// The tasks will be handled in `handleResolveLockTasks` goroutine.
-	resolveLockTaskCh chan resolveLockTask
-	// failureBuffer serializes region failures before recovery.
-	failureBuffer *failureBuffer
+	if s.regionScheduler == nil {
+		s.regionScheduler = newRegionRequestScheduler(s)
+	}
+	if s.subscribedSpans == nil {
+		s.subscribedSpans = newSubscribedSpanSet(s)
+	}
+	if s.staleLockResolver == nil {
+		s.staleLockResolver = newStaleLockResolver(s)
+	}
 }
 
 func (s *subscriptionClient) ensureRegionRuntime(region *regionInfo, now time.Time) {
@@ -324,14 +319,9 @@ func NewSubscriptionClient(
 		regionRuntimeRegistry: newRegionRuntimeRegistry(),
 
 		credential: credential,
-
-		rangeTaskCh:       make(chan rangeTask, 1024),
-		regionTaskQueue:   NewPriorityQueue(),
-		resolveLockTaskCh: make(chan resolveLockTask, 1024),
-		failureBuffer:     newFailureBuffer(),
 	}
 	subClient.ctx, subClient.cancel = context.WithCancel(context.Background())
-	subClient.totalSpans.spanMap = make(map[SubscriptionID]*subscribedSpan)
+	subClient.ensureHelpers()
 
 	option := dynstream.NewOption()
 	// Note: it is max batch size of the kv sent from tikv(not committed rows)
@@ -368,7 +358,7 @@ func (s *subscriptionClient) updateMetrics(ctx context.Context) error {
 		case <-ctx.Done():
 			return ctx.Err()
 		case <-ticker.C:
-			resolvedTsLag := s.GetResolvedTsLag()
+			resolvedTsLag := s.subscribedSpans.getResolvedTsLag()
 			if resolvedTsLag > 0 {
 				metrics.LogPullerResolvedTsLag.Set(resolvedTsLag)
 			}
@@ -394,7 +384,7 @@ func (s *subscriptionClient) updateMetrics(ctx context.Context) error {
 				).Set(float64(areaMetric.MemoryUsage()))
 			}
 
-			pendingRegionReqCount := s.pendingRequestCount()
+			pendingRegionReqCount := s.requestedStores.pendingRequestCount()
 			metrics.SubscriptionClientRequestedRegionCount.WithLabelValues("pending").Set(float64(pendingRegionReqCount))
 
 			if counts := s.regionRuntimePhaseCounts(); counts != nil {
@@ -403,21 +393,14 @@ func (s *subscriptionClient) updateMetrics(ctx context.Context) error {
 				}
 			}
 
-			count := 0
-			s.totalSpans.RLock()
-			for _, rt := range s.totalSpans.spanMap {
-				count += rt.rangeLock.Len()
-			}
-			s.totalSpans.RUnlock()
-			metrics.SubscriptionClientSubscribedRegionCount.Set(float64(count))
+			metrics.SubscriptionClientSubscribedRegionCount.Set(float64(s.subscribedSpans.requestedRegionCount()))
 		}
 	}
 }
 
 // Subscribe the given table span.
 // NOTE: `span.TableID` must be set correctly.
-// It new a subscribedSpan and store it in `s.totalSpans`,
-// and send a rangeTask to `s.rangeTaskCh`.
+// It creates a subscribed span, registers it, and schedules the bootstrap range request.
 func (s *subscriptionClient) Subscribe(
 	subID SubscriptionID,
 	span heartbeatpb.TableSpan,
@@ -432,36 +415,32 @@ func (s *subscriptionClient) Subscribe(
 		return
 	}
 
-	rt := s.newSubscribedSpan(subID, span, startTs, consumeKVEvents, advanceResolvedTs, advanceInterval, bdrMode)
-	s.totalSpans.Lock()
-	s.totalSpans.spanMap[subID] = rt
-	s.totalSpans.Unlock()
+	s.ensureHelpers()
+	rt := s.subscribedSpans.newSubscribedSpan(subID, span, startTs, consumeKVEvents, advanceResolvedTs, advanceInterval, bdrMode)
+	s.subscribedSpans.add(subID, rt)
 
 	areaSetting := dynstream.NewAreaSettingsWithMaxPendingSize(1*1024*1024*1024, dynstream.MemoryControlForPuller, "logPuller") // 1GB
 	s.ds.AddPath(rt.subID, rt, areaSetting)
 
-	select {
-	case <-s.ctx.Done():
+	if !s.regionScheduler.scheduleRangeRequest(s.ctx, span, rt, rt.filterLoop, TaskLowPrior) {
 		log.Warn("subscribes span failed, the subscription client has closed")
-	case s.rangeTaskCh <- rangeTask{span: span, subscribedSpan: rt, filterLoop: rt.filterLoop, priority: TaskLowPrior}:
-		log.Info("subscribes span done", zap.Uint64("subscriptionID", uint64(subID)),
-			zap.Int64("tableID", span.TableID), zap.Uint64("startTs", startTs),
-			zap.String("startKey", spanz.HexKey(span.StartKey)), zap.String("endKey", spanz.HexKey(span.EndKey)))
+		return
 	}
+	log.Info("subscribes span done", zap.Uint64("subscriptionID", uint64(subID)),
+		zap.Int64("tableID", span.TableID), zap.Uint64("startTs", startTs),
+		zap.String("startKey", spanz.HexKey(span.StartKey)), zap.String("endKey", spanz.HexKey(span.EndKey)))
 }
 
 // Unsubscribe the given table span. All covered regions will be deregistered asynchronously.
 // NOTE: `span.TableID` must be set correctly.
 func (s *subscriptionClient) Unsubscribe(subID SubscriptionID) {
-	// NOTE: `subID` is cleared from `s.totalSpans` in `onTableDrained`.
-	s.totalSpans.Lock()
-	rt := s.totalSpans.spanMap[subID]
-	s.totalSpans.Unlock()
+	s.ensureHelpers()
+	rt := s.subscribedSpans.get(subID)
 	if rt == nil {
 		log.Warn("unknown subscription", zap.Uint64("subscriptionID", uint64(subID)))
 		return
 	}
-	s.setTableStopped(rt)
+	s.subscribedSpans.setTableStopped(rt)
 
 	log.Info("unsubscribe span success",
 		zap.Uint64("subscriptionID", uint64(rt.subID)),
@@ -524,19 +503,20 @@ func (s *subscriptionClient) Run(ctx context.Context) error {
 		log.Warn("subscription client should be in test mode, skip run")
 		return nil
 	}
+	s.ensureHelpers()
 	s.clusterID = s.pd.GetClusterID(ctx)
 
 	g, ctx := errgroup.WithContext(ctx)
 
 	g.Go(func() error { return s.updateMetrics(ctx) })
 	g.Go(func() error { return s.handleDSFeedBack(ctx) })
-	g.Go(func() error { return s.handleRangeTasks(ctx) })
-	g.Go(func() error { return s.handleRegions(ctx, g) })
-	g.Go(func() error { return s.handleFailures(ctx) })
-	g.Go(func() error { return s.runResolveLockChecker(ctx) })
-	g.Go(func() error { return s.handleResolveLockTasks(ctx) })
+	g.Go(func() error { return s.regionScheduler.handleRangeTasks(ctx) })
+	g.Go(func() error { return s.regionScheduler.handleRegions(ctx, g) })
+	g.Go(func() error { return s.regionScheduler.handleFailures(ctx) })
+	g.Go(func() error { return s.staleLockResolver.runResolveLockChecker(ctx) })
+	g.Go(func() error { return s.staleLockResolver.handleResolveLockTasks(ctx) })
 	g.Go(func() error { return s.logSlowRegions(ctx) })
-	g.Go(func() error { return s.failureBuffer.run(ctx) })
+	g.Go(func() error { return s.regionScheduler.runFailureBuffer(ctx) })
 
 	log.Info("subscription client starts")
 	defer log.Info("subscription client exits")
@@ -551,331 +531,44 @@ func (s *subscriptionClient) Close(ctx context.Context) error {
 	s.cond.Broadcast()
 	s.mu.Unlock()
 	s.ds.Close()
-	s.regionTaskQueue.Close()
+	if s.regionScheduler != nil {
+		s.regionScheduler.close()
+	}
 	return nil
 }
 
 func (s *subscriptionClient) setTableStopped(rt *subscribedSpan) {
-	log.Info("subscription client starts to stop table",
-		zap.Uint64("subscriptionID", uint64(rt.subID)))
-
-	// Set stopped to true so we can stop handling region events from the table.
-	// Then send a special singleRegionInfo so every store worker deregisters the table.
-	if rt.stopped.CompareAndSwap(false, true) {
-		s.regionTaskQueue.Push(NewRegionPriorityTask(
-			TaskHighPrior,
-			regionInfo{subscribedSpan: rt, filterLoop: rt.filterLoop},
-			s.pdClock.CurrentTS(),
-		))
-		if rt.rangeLock.Stop() {
-			s.onTableDrained(rt)
-		}
-	}
+	s.ensureHelpers()
+	s.subscribedSpans.setTableStopped(rt)
 }
 
-func (s *subscriptionClient) onTableDrained(rt *subscribedSpan) {
-	log.Info("subscription client stop span is finished",
-		zap.Uint64("subscriptionID", uint64(rt.subID)))
-
-	s.removeSubscriptionRuntime(rt.subID)
-
-	err := s.ds.RemovePath(rt.subID)
-	if err != nil {
-		log.Warn("subscription client remove path failed",
-			zap.Uint64("subscriptionID", uint64(rt.subID)),
-			zap.Error(err))
-	}
-	s.totalSpans.Lock()
-	defer s.totalSpans.Unlock()
-	delete(s.totalSpans.spanMap, rt.subID)
+func (s *subscriptionClient) newSubscribedSpan(
+	subID SubscriptionID,
+	span heartbeatpb.TableSpan,
+	startTs uint64,
+	consumeKVEvents func(raw []common.RawKVEntry, wakeCallback func()) bool,
+	advanceResolvedTs func(ts uint64),
+	advanceInterval int64,
+	filterLoop bool,
+) *subscribedSpan {
+	s.ensureHelpers()
+	return s.subscribedSpans.newSubscribedSpan(
+		subID, span, startTs, consumeKVEvents, advanceResolvedTs, advanceInterval, filterLoop)
 }
 
-func (s *subscriptionClient) pendingRequestCount() int {
-	pendingRegionReqCount := 0
-	s.stores.Range(func(_, value any) bool {
-		store := value.(*requestedStore)
-		for _, worker := range store.snapshotWorkers() {
-			pendingRegionReqCount += worker.requestCache.getPendingCount()
-		}
-		return true
-	})
-	return pendingRegionReqCount
-}
-
-func (s *subscriptionClient) cleanupRequestedStores() {
-	s.stores.Range(func(_, value any) bool {
-		store := value.(*requestedStore)
-		for _, worker := range store.snapshotWorkers() {
-			worker.requestCache.clear()
-		}
-		return true
-	})
-}
-
-func (s *subscriptionClient) perWorkerQueueSize() int {
-	config := config.GetGlobalServerConfig()
-	perWorkerQueueSize := config.Debug.Puller.PendingRegionRequestQueueSize / int(s.config.RegionRequestWorkerPerStore)
-	if perWorkerQueueSize <= 0 {
-		log.Warn("pending region request queue size is smaller than the number of workers, adjust per worker queue size to 1",
-			zap.Int("pendingRegionRequestQueueSize", config.Debug.Puller.PendingRegionRequestQueueSize),
-			zap.Uint("regionRequestWorkerPerStore", s.config.RegionRequestWorkerPerStore))
-		return 1
-	}
-	return perWorkerQueueSize
-}
-
-func (s *subscriptionClient) getOrCreateRequestedStore(
-	ctx context.Context,
-	eg *errgroup.Group,
-	storeAddr string,
-) *requestedStore {
-	if value, ok := s.stores.Load(storeAddr); ok {
-		return value.(*requestedStore)
-	}
-
-	store := newRequestedStore(storeAddr)
-	s.stores.Store(storeAddr, store)
-
-	perWorkerQueueSize := s.perWorkerQueueSize()
-	for i := uint(0); i < s.config.RegionRequestWorkerPerStore; i++ {
-		store.addWorker(newRegionRequestWorker(ctx, s, s.credential, eg, store, perWorkerQueueSize))
-	}
-	return store
+func (s *subscriptionClient) GetResolvedTsLag() float64 {
+	s.ensureHelpers()
+	return s.subscribedSpans.getResolvedTsLag()
 }
 
 func (s *subscriptionClient) enqueueRegionToAllStores(ctx context.Context, region regionInfo) (bool, error) {
-	enqueued := true
-	var firstErr error
-	s.stores.Range(func(_ any, value any) bool {
-		store := value.(*requestedStore)
-		for _, worker := range store.snapshotWorkers() {
-			ok, err := worker.add(ctx, region, true)
-			if err != nil {
-				firstErr = err
-				enqueued = false
-				return false
-			}
-			if !ok {
-				enqueued = false
-				// It is likely the store is busy, no need to try other workers in this store now.
-				break
-			}
-		}
-		return true
-	})
-	return enqueued, firstErr
-}
-
-func (s *subscriptionClient) attachRPCContextForRegion(ctx context.Context, region regionInfo) (regionInfo, bool) {
-	bo := tikv.NewBackoffer(ctx, tikvRequestMaxBackoff)
-	rpcCtx, err := s.regionCache.GetTiKVRPCContext(bo, region.verID, kvclientv2.ReplicaReadLeader, 0)
-	if rpcCtx != nil {
-		region.rpcCtx = rpcCtx
-		return region, true
-	}
-	if err != nil {
-		log.Debug("subscription client get rpc context fail",
-			zap.Uint64("subscriptionID", uint64(region.subscribedSpan.subID)),
-			zap.Uint64("regionID", region.verID.GetID()),
-			zap.Error(err))
-	}
-	s.submitDirectFailure(newRPCCtxUnavailableFailure(region))
-	return region, false
-}
-
-// handleRegions receives regionInfo from regionTaskQueue, attaches rpcCtx to them,
-// then sends them to the corresponding requestedStore.
-func (s *subscriptionClient) handleRegions(ctx context.Context, eg *errgroup.Group) error {
-	defer s.cleanupRequestedStores()
-
-	for {
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		default:
-		}
-
-		regionTask, err := s.regionTaskQueue.Pop(ctx)
-		if err != nil {
-			return err
-		}
-
-		region := regionTask.GetRegionInfo()
-		if region.isStopped() {
-			enqueued, err := s.enqueueRegionToAllStores(ctx, region)
-			if err != nil {
-				return err
-			}
-			if !enqueued {
-				log.Debug("enqueue stop request failed, retry later",
-					zap.Uint64("subscriptionID", uint64(region.subscribedSpan.subID)))
-				s.regionTaskQueue.Push(regionTask)
-			}
-			continue
-		}
-
-		region, ok := s.attachRPCContextForRegion(ctx, region)
-		if !ok {
-			continue
-		}
-		s.updateRegionRuntimeInfo(region)
-		s.markRegionRPCReady(region, time.Now())
-
-		store := s.getOrCreateRequestedStore(ctx, eg, region.rpcCtx.Addr)
-		worker := store.getRequestWorker()
-		force := regionTask.Priority() <= forcedPriorityBase
-
-		ok, err = worker.add(ctx, region, force)
-		if err != nil {
-			log.Warn("subscription client add region request failed",
-				zap.Uint64("subscriptionID", uint64(region.subscribedSpan.subID)),
-				zap.Uint64("regionID", region.verID.GetID()),
-				zap.Error(err))
-			return err
-		}
-		if !ok {
-			s.regionTaskQueue.Push(regionTask)
-			continue
-		}
-
-		log.Debug("subscription client will request a region",
-			zap.Uint64("workID", worker.workerID),
-			zap.Uint64("subscriptionID", uint64(region.subscribedSpan.subID)),
-			zap.Uint64("regionID", region.verID.GetID()),
-			zap.String("addr", store.storeAddr))
-	}
-}
-
-func (s *subscriptionClient) handleRangeTasks(ctx context.Context) error {
-	g, ctx := errgroup.WithContext(ctx)
-	// Limit the concurrent number of goroutines to convert range tasks to region tasks.
-	g.SetLimit(1024)
-	for {
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		case task := <-s.rangeTaskCh:
-			g.Go(func() error {
-				return s.divideSpanAndScheduleRegionRequests(ctx, task.span, task.subscribedSpan, task.filterLoop, task.priority)
-			})
-		}
-	}
-}
-
-// divideSpanAndScheduleRegionRequests processes the specified span by dividing it into
-// manageable regions and schedules requests to subscribe to these regions.
-// 1. Load regions from PD.
-// 2. Find the intersection of each region.span and the subscribedSpan.span.
-// 3. Schedule a region request to subscribe the region.
-func (s *subscriptionClient) divideSpanAndScheduleRegionRequests(
-	ctx context.Context,
-	span heartbeatpb.TableSpan,
-	subscribedSpan *subscribedSpan,
-	filterLoop bool,
-	taskType TaskType,
-) error {
-	// Limit the number of regions loaded at a time to make the load more stable.
-	limit := 1024
-	nextSpan := span
-	backoffBeforeLoad := false
-	for {
-		if backoffBeforeLoad {
-			if err := util.Hang(ctx, loadRegionRetryInterval); err != nil {
-				return err
-			}
-			backoffBeforeLoad = false
-		}
-		log.Debug("subscription client is going to load regions",
-			zap.Uint64("subscriptionID", uint64(subscribedSpan.subID)),
-			zap.Any("span", common.FormatTableSpan(&nextSpan)))
-
-		backoff := tikv.NewBackoffer(ctx, tikvRequestMaxBackoff)
-		regions, err := s.regionCache.BatchLoadRegionsWithKeyRange(backoff, nextSpan.StartKey, nextSpan.EndKey, limit)
-		if err != nil {
-			log.Warn("subscription client load regions failed",
-				zap.Uint64("subscriptionID", uint64(subscribedSpan.subID)),
-				zap.Any("span", common.FormatTableSpan(&nextSpan)),
-				zap.Error(err))
-			backoffBeforeLoad = true
-			continue
-		}
-		regionMetas := make([]*metapb.Region, 0, len(regions))
-		for _, region := range regions {
-			if meta := region.GetMeta(); meta != nil {
-				regionMetas = append(regionMetas, meta)
-			}
-		}
-		regionMetas = regionlock.CutRegionsLeftCoverSpan(regionMetas, nextSpan)
-		if len(regionMetas) == 0 {
-			log.Warn("subscription client load regions with holes",
-				zap.Uint64("subscriptionID", uint64(subscribedSpan.subID)),
-				zap.Any("span", common.FormatTableSpan(&nextSpan)))
-			backoffBeforeLoad = true
-			continue
-		}
-
-		for _, regionMeta := range regionMetas {
-			regionSpan := heartbeatpb.TableSpan{
-				StartKey:   regionMeta.StartKey,
-				EndKey:     regionMeta.EndKey,
-				KeyspaceID: subscribedSpan.span.KeyspaceID,
-			}
-			// NOTE: the End key return by the PD API will be nil to represent the biggest key.
-			// So we need to fix it by calling spanz.HackSpan.
-			regionSpan = common.HackTableSpan(regionSpan)
-
-			// Find the intersection of the regionSpan returned by PD and the subscribedSpan.span.
-			// The intersection is the span that needs to be subscribed.
-			intersectSpan := common.GetIntersectSpan(subscribedSpan.span, regionSpan)
-			if common.IsEmptySpan(intersectSpan) {
-				log.Panic("subscription client check spans intersect shouldn't fail",
-					zap.Uint64("subscriptionID", uint64(subscribedSpan.subID)))
-			}
-
-			regionInfo := newRegionInfo(
-				tikv.NewRegionVerID(regionMeta.Id, regionMeta.RegionEpoch.ConfVer, regionMeta.RegionEpoch.Version),
-				intersectSpan,
-				nil,
-				subscribedSpan,
-				filterLoop,
-			)
-			s.scheduleRegionRequest(ctx, regionInfo, taskType)
-
-			nextSpan.StartKey = regionMeta.EndKey
-			// If the nextSpan.StartKey is larger than the subscribedSpan.span.EndKey,
-			// it means all span of the subscribedSpan have been requested. So we return.
-			if common.EndCompare(nextSpan.StartKey, span.EndKey) >= 0 {
-				return nil
-			}
-		}
-	}
+	s.ensureHelpers()
+	return s.requestedStores.enqueueRegionToAllStores(ctx, region)
 }
 
 func (s *subscriptionClient) scheduleRegionRequest(ctx context.Context, region regionInfo, priority TaskType) {
-	s.ensureRegionRuntime(&region, time.Now())
-	lockRangeResult := region.subscribedSpan.rangeLock.LockRange(
-		ctx, region.span.StartKey, region.span.EndKey, region.verID.GetID(), region.verID.GetVer())
-
-	if lockRangeResult.Status == regionlock.LockRangeStatusWait {
-		s.transitionRegionRuntime(region, regionPhaseRangeLockWait, time.Now())
-		lockRangeResult = lockRangeResult.WaitFn()
-	}
-
-	switch lockRangeResult.Status {
-	case regionlock.LockRangeStatusSuccess:
-		region.lockedRangeState = lockRangeResult.LockedRangeState
-		s.markRegionQueued(region, lockRangeResult.LockedRangeState.Created, time.Now())
-		s.regionTaskQueue.Push(NewRegionPriorityTask(priority, region, s.pdClock.CurrentTS()))
-	case regionlock.LockRangeStatusStale:
-		s.removeRegionRuntime(region, time.Now())
-		for _, retryRange := range lockRangeResult.RetryRanges {
-			s.scheduleRangeRequest(ctx, retryRange, region.subscribedSpan, region.filterLoop, priority)
-		}
-	case regionlock.LockRangeStatusCancel:
-		s.removeRegionRuntime(region, time.Now())
-	default:
-		return
-	}
+	s.ensureHelpers()
+	s.regionScheduler.scheduleRegionRequest(ctx, region, priority)
 }
 
 func (s *subscriptionClient) scheduleRangeRequest(
@@ -885,93 +578,18 @@ func (s *subscriptionClient) scheduleRangeRequest(
 	filterLoop bool,
 	priority TaskType,
 ) {
-	select {
-	case <-ctx.Done():
-	case s.rangeTaskCh <- rangeTask{
-		span:           span,
-		subscribedSpan: subscribedSpan,
-		filterLoop:     filterLoop,
-		priority:       priority,
-	}:
-	}
-}
-
-type failureBuffer struct {
-	sync.Mutex
-	pending []regionFailureInfo
-	ch      chan regionFailureInfo
-	notify  chan struct{}
-}
-
-func newFailureBuffer() *failureBuffer {
-	return &failureBuffer{
-		pending: make([]regionFailureInfo, 0, 1024),
-		ch:      make(chan regionFailureInfo, 1024),
-		notify:  make(chan struct{}, 1024),
-	}
-}
-
-func (b *failureBuffer) enqueue(failure regionFailureInfo) {
-	b.Lock()
-	defer b.Unlock()
-	b.pending = append(b.pending, failure)
-	select {
-	case b.notify <- struct{}{}:
-	default:
-	}
-}
-
-func (b *failureBuffer) run(ctx context.Context) error {
-	ticker := time.NewTicker(10 * time.Millisecond)
-	defer ticker.Stop()
-
-	dispatchOne := func() {
-		b.Lock()
-		if len(b.pending) == 0 {
-			b.Unlock()
-			return
-		}
-		failure := b.pending[0]
-		b.pending = b.pending[1:]
-		b.Unlock()
-
-		select {
-		case <-ctx.Done():
-			log.Info("subscription client dispatch failure buffer done")
-		case b.ch <- failure:
-		}
-	}
-
-	for {
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		case <-ticker.C:
-			dispatchOne()
-		case <-b.notify:
-			dispatchOne()
-		}
-	}
+	s.ensureHelpers()
+	s.regionScheduler.scheduleRangeRequest(ctx, span, subscribedSpan, filterLoop, priority)
 }
 
 func (s *subscriptionClient) submitDirectFailure(failure regionFailureInfo) {
-	s.recordRegionRuntimeError(failure.regionInfo, failure.err, time.Now())
-	if failure.subscribedSpan.rangeLock.UnlockRange(
-		failure.span.StartKey, failure.span.EndKey,
-		failure.verID.GetID(), failure.verID.GetVer(), failure.resolvedTs()) {
-		s.onTableDrained(failure.subscribedSpan)
-		return
-	}
-	s.failureBuffer.enqueue(failure)
+	s.ensureHelpers()
+	s.regionScheduler.submitDirectFailure(failure)
 }
 
 func (s *subscriptionClient) submitOrderedFailure(state *regionFeedState) (regionFailureInfo, bool) {
-	failure, removed := state.detachFailure()
-	if !removed {
-		return regionFailureInfo{}, false
-	}
-	s.submitDirectFailure(failure)
-	return failure, true
+	s.ensureHelpers()
+	return s.regionScheduler.submitOrderedFailure(state)
 }
 
 func (s *subscriptionClient) submitWorkerSessionFailure(
@@ -979,283 +597,13 @@ func (s *subscriptionClient) submitWorkerSessionFailure(
 	pendingRegions []regionInfo,
 	sessionFailure workerSessionFailure,
 ) {
-	for subID, states := range startedRegions {
-		for _, state := range states {
-			state.markStopped(normalizeWorkerSessionFailure(state.getRegionInfo(), sessionFailure))
-			s.pushRegionEventToDS(subID, regionEvent{
-				states: []*regionFeedState{state},
-			})
-		}
-	}
-
-	for _, region := range pendingRegions {
-		if region.isStopped() {
-			continue
-		}
-		s.submitDirectFailure(normalizeWorkerSessionFailure(region, sessionFailure))
-	}
-}
-
-func (s *subscriptionClient) handleFailures(ctx context.Context) error {
-	for {
-		select {
-		case <-ctx.Done():
-			log.Info("subscription client handle failures and exit")
-			return ctx.Err()
-		case failure := <-s.failureBuffer.ch:
-			if err := s.handleFailure(ctx, failure); err != nil {
-				return err
-			}
-		}
-	}
-}
-
-func (s *subscriptionClient) retryRegion(ctx context.Context, failure regionFailureInfo, priority TaskType) {
-	s.markRegionRetryPending(failure.regionInfo, failure.err, time.Now())
-	s.scheduleRegionRequest(ctx, failure.regionInfo, priority)
-}
-
-func (s *subscriptionClient) reloadRegionRange(ctx context.Context, failure regionFailureInfo) {
-	s.removeRegionRuntime(failure.regionInfo, time.Now())
-	s.scheduleRangeRequest(ctx, failure.span, failure.subscribedSpan, failure.filterLoop, TaskHighPrior)
-}
-
-func (s *subscriptionClient) handleRegionFailure(ctx context.Context, failure regionFailureInfo) error {
-	switch failure.kind {
-	case regionFailureKindTiKVEvent:
-		eventErr, ok := failure.err.(*eventError)
-		if !ok || eventErr == nil {
-			return errors.New("invalid tikv event failure")
-		}
-
-		innerErr := eventErr.err
-		if notLeader := innerErr.GetNotLeader(); notLeader != nil {
-			metricFeedNotLeaderCounter.Inc()
-			if s.regionCache != nil {
-				s.regionCache.UpdateLeader(failure.verID, notLeader.GetLeader(), failure.rpcCtx.AccessIdx)
-			}
-			s.retryRegion(ctx, failure, TaskHighPrior)
-			return nil
-		}
-		if innerErr.GetEpochNotMatch() != nil {
-			metricFeedEpochNotMatchCounter.Inc()
-			s.reloadRegionRange(ctx, failure)
-			return nil
-		}
-		if innerErr.GetRegionNotFound() != nil {
-			metricFeedRegionNotFoundCounter.Inc()
-			s.reloadRegionRange(ctx, failure)
-			return nil
-		}
-		if innerErr.GetCongested() != nil {
-			metricKvCongestedCounter.Inc()
-			s.retryRegion(ctx, failure, TaskLowPrior)
-			return nil
-		}
-		if innerErr.GetServerIsBusy() != nil {
-			metricKvIsBusyCounter.Inc()
-			s.retryRegion(ctx, failure, TaskLowPrior)
-			return nil
-		}
-		if duplicated := innerErr.GetDuplicateRequest(); duplicated != nil {
-			metricFeedDuplicateRequestCounter.Inc()
-			return errors.New("duplicate request")
-		}
-		if compatibility := innerErr.GetCompatibility(); compatibility != nil {
-			return cerror.ErrVersionIncompatible.GenWithStackByArgs(compatibility)
-		}
-		if mismatch := innerErr.GetClusterIdMismatch(); mismatch != nil {
-			return cerror.ErrClusterIDMismatch.GenWithStackByArgs(mismatch.Current, mismatch.Request)
-		}
-
-		log.Warn("empty or unknown cdc error",
-			zap.Uint64("subscriptionID", uint64(failure.subscribedSpan.subID)),
-			zap.Stringer("error", innerErr))
-		metricFeedUnknownErrorCounter.Inc()
-		s.retryRegion(ctx, failure, TaskHighPrior)
-		return nil
-	case regionFailureKindRPCCtxUnavailable:
-		metricFeedRPCCtxUnavailable.Inc()
-		s.reloadRegionRange(ctx, failure)
-		return nil
-	default:
-		return errors.New("unexpected region failure kind")
-	}
-}
-
-func (s *subscriptionClient) handleStoreSessionFailure(ctx context.Context, failure regionFailureInfo) error {
-	switch failure.kind {
-	case regionFailureKindGetStore:
-		metricGetStoreErr.Inc()
-		if s.regionCache != nil {
-			bo := tikv.NewBackoffer(ctx, tikvRequestMaxBackoff)
-			s.regionCache.OnSendFail(bo, failure.rpcCtx, true, errors.Cause(failure.err))
-		}
-		s.reloadRegionRange(ctx, failure)
-		return nil
-	case regionFailureKindSendRequestToStore:
-		metricStoreSendRequestErr.Inc()
-		if s.regionCache != nil {
-			bo := tikv.NewBackoffer(ctx, tikvRequestMaxBackoff)
-			s.regionCache.OnSendFail(bo, failure.rpcCtx, regionScheduleReload, errors.Cause(failure.err))
-		}
-		s.retryRegion(ctx, failure, TaskHighPrior)
-		return nil
-	default:
-		return errors.New("unexpected store session failure kind")
-	}
-}
-
-func (s *subscriptionClient) handleSubscriptionFailure(failure regionFailureInfo) error {
-	switch failure.kind {
-	case regionFailureKindRequestCancelled, regionFailureKindSubscriptionStopped:
-		s.removeRegionRuntime(failure.regionInfo, time.Now())
-		return nil
-	default:
-		return errors.New("unexpected subscription failure kind")
-	}
+	s.ensureHelpers()
+	s.regionScheduler.submitWorkerSessionFailure(startedRegions, pendingRegions, sessionFailure)
 }
 
 func (s *subscriptionClient) handleFailure(ctx context.Context, failure regionFailureInfo) error {
-	log.Debug("cdc region failure",
-		zap.Uint64("subscriptionID", uint64(failure.subscribedSpan.subID)),
-		zap.Uint64("regionID", failure.verID.GetID()),
-		zap.Stringer("failureScope", failure.scope),
-		zap.Stringer("failureSource", failure.source),
-		zap.Stringer("failureKind", failure.kind),
-		zap.Error(failure.err))
-
-	var err error
-	switch failure.scope {
-	case regionFailureScopeRegion:
-		err = s.handleRegionFailure(ctx, failure)
-	case regionFailureScopeStoreSession:
-		err = s.handleStoreSessionFailure(ctx, failure)
-	case regionFailureScopeSubscription:
-		err = s.handleSubscriptionFailure(failure)
-	default:
-		err = errors.New("unknown failure scope")
-	}
-	if err == nil {
-		return nil
-	}
-
-	log.Warn("subscription client meets an internal error, fail the changefeed",
-		zap.Uint64("subscriptionID", uint64(failure.subscribedSpan.subID)),
-		zap.Stringer("failureScope", failure.scope),
-		zap.Stringer("failureSource", failure.source),
-		zap.Stringer("failureKind", failure.kind),
-		zap.Error(err))
-	return err
-}
-
-type subscriptionAndTargetTs struct {
-	subSpan  *subscribedSpan
-	targetTs uint64
-}
-
-func (s *subscriptionClient) runResolveLockChecker(ctx context.Context) error {
-	resolveLockTicker := time.NewTicker(resolveLockTickInterval)
-	defer resolveLockTicker.Stop()
-	maxCacheSize := 1024
-	subSpanAndTsCache := make([]subscriptionAndTargetTs, 0, maxCacheSize)
-	// getResolvedTargetTs returns the targetTs to resolve stale locks. 0 means no need to resolve.
-	getResolvedTargetTs := func(subSpan *subscribedSpan, currentTime time.Time) uint64 {
-		resolvedTsUpdated := time.Unix(subSpan.resolvedTsUpdated.Load(), 0)
-		if !subSpan.initialized.Load() || time.Since(resolvedTsUpdated) < resolveLockFence {
-			return 0
-		}
-		resolvedTs := subSpan.resolvedTs.Load()
-		resolvedTime := oracle.GetTimeFromTS(resolvedTs)
-		if currentTime.Sub(resolvedTime) < resolveLockFence {
-			return 0
-		}
-		return oracle.GoTimeToTS(resolvedTime.Add(resolveLockFence))
-	}
-
-	for {
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		case <-resolveLockTicker.C:
-		}
-		currentTime := s.pdClock.CurrentTime()
-		s.totalSpans.Lock()
-		for _, subSpan := range s.totalSpans.spanMap {
-			if subSpan != nil {
-				targetTs := getResolvedTargetTs(subSpan, currentTime)
-				if targetTs > 0 {
-					subSpanAndTsCache = append(subSpanAndTsCache, subscriptionAndTargetTs{
-						subSpan:  subSpan,
-						targetTs: targetTs,
-					})
-				}
-			}
-		}
-		s.totalSpans.Unlock()
-		for _, subSpanAndTs := range subSpanAndTsCache {
-			subSpanAndTs.subSpan.resolveStaleLocks(subSpanAndTs.targetTs)
-		}
-		subSpanAndTsCache = subSpanAndTsCache[:0]
-		if cap(subSpanAndTsCache) > maxCacheSize {
-			subSpanAndTsCache = make([]subscriptionAndTargetTs, 0, maxCacheSize)
-		}
-	}
-}
-
-func gcResolveLastRunMap(resolveLastRun map[uint64]time.Time, now time.Time) map[uint64]time.Time {
-	if len(resolveLastRun) <= resolveLastRunGCThreshold {
-		return resolveLastRun
-	}
-
-	copied := make(map[uint64]time.Time, len(resolveLastRun))
-	for regionID, lastRun := range resolveLastRun {
-		if now.Sub(lastRun) < resolveLockMinInterval {
-			copied[regionID] = lastRun
-		}
-	}
-	return copied
-}
-
-func (s *subscriptionClient) handleResolveLockTasks(ctx context.Context) error {
-	resolveLastRun := make(map[uint64]time.Time)
-
-	doResolve := func(keyspaceID uint32, regionID uint64, state *regionlock.LockedRangeState, targetTs uint64) {
-		if state.ResolvedTs.Load() > targetTs || !state.Initialized.Load() {
-			return
-		}
-
-		lastRun, ok := resolveLastRun[regionID]
-		if ok {
-			if time.Since(lastRun) < resolveLockMinInterval {
-				return
-			}
-		}
-
-		if err := s.lockResolver.Resolve(ctx, keyspaceID, regionID, targetTs); err != nil {
-			log.Warn("subscription client resolve lock fail",
-				zap.Uint32("keyspaceID", keyspaceID),
-				zap.Uint64("regionID", regionID),
-				zap.Uint64("targetTs", targetTs),
-				zap.Time("lastRun", lastRun),
-				zap.Any("state", state),
-				zap.Error(err))
-		}
-		resolveLastRun[regionID] = time.Now()
-	}
-
-	gcTicker := time.NewTicker(resolveLockMinInterval * 3 / 2)
-	defer gcTicker.Stop()
-	for {
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		case <-gcTicker.C:
-			resolveLastRun = gcResolveLastRunMap(resolveLastRun, time.Now())
-		case task := <-s.resolveLockTaskCh:
-			doResolve(task.keyspaceID, task.regionID, task.state, task.targetTs)
-		}
-	}
+	s.ensureHelpers()
+	return s.regionScheduler.handleFailure(ctx, failure)
 }
 
 func (s *subscriptionClient) logSlowRegions(ctx context.Context) error {
@@ -1269,107 +617,32 @@ func (s *subscriptionClient) logSlowRegions(ctx context.Context) error {
 		}
 
 		currTime := s.pdClock.CurrentTime()
-		s.totalSpans.RLock()
 		slowInitializeRegion := 0
-		for subscriptionID, rt := range s.totalSpans.spanMap {
+		for _, entry := range s.subscribedSpans.snapshot() {
+			rt := entry.span
 			attr := rt.rangeLock.IterAll(nil)
 			ckptTime := oracle.GetTimeFromTS(attr.SlowestRegion.ResolvedTs)
 			if attr.SlowestRegion.Initialized {
 				if currTime.Sub(ckptTime) > 6*resolveLockMinInterval {
 					log.Info("subscription client finds a initialized slow region",
-						zap.Uint64("subscriptionID", uint64(subscriptionID)),
+						zap.Uint64("subscriptionID", uint64(entry.subID)),
 						zap.Any("slowRegion", attr.SlowestRegion))
 				}
 			} else if currTime.Sub(attr.SlowestRegion.Created) > 10*time.Minute {
 				slowInitializeRegion += 1
 				log.Info("subscription client initializes a region too slow",
-					zap.Uint64("subscriptionID", uint64(subscriptionID)),
+					zap.Uint64("subscriptionID", uint64(entry.subID)),
 					zap.Any("slowRegion", attr.SlowestRegion))
 			} else if currTime.Sub(ckptTime) > 10*time.Minute {
 				log.Info("subscription client finds a uninitialized slow region",
-					zap.Uint64("subscriptionID", uint64(subscriptionID)),
+					zap.Uint64("subscriptionID", uint64(entry.subID)),
 					zap.Any("slowRegion", attr.SlowestRegion))
 			}
 			if len(attr.UnLockedRanges) > 0 {
 				log.Info("subscription client holes exist",
-					zap.Uint64("subscriptionID", uint64(subscriptionID)),
+					zap.Uint64("subscriptionID", uint64(entry.subID)),
 					zap.Any("holes", attr.UnLockedRanges))
 			}
 		}
-		s.totalSpans.RUnlock()
 	}
-}
-
-func (s *subscriptionClient) newSubscribedSpan(
-	subID SubscriptionID,
-	span heartbeatpb.TableSpan,
-	startTs uint64,
-	consumeKVEvents func(raw []common.RawKVEntry, wakeCallback func()) bool,
-	advanceResolvedTs func(ts uint64),
-	advanceInterval int64,
-	filterLoop bool,
-) *subscribedSpan {
-	rangeLock := regionlock.NewRangeLock(uint64(subID), span.StartKey, span.EndKey, startTs)
-
-	rt := &subscribedSpan{
-		subID:      subID,
-		span:       span,
-		startTs:    startTs,
-		filterLoop: filterLoop,
-		rangeLock:  rangeLock,
-
-		consumeKVEvents:   consumeKVEvents,
-		advanceResolvedTs: advanceResolvedTs,
-		advanceInterval:   advanceInterval,
-	}
-	rt.initialized.Store(false)
-	rt.resolvedTsUpdated.Store(time.Now().Unix())
-	rt.resolvedTs.Store(startTs)
-
-	rt.tryResolveLock = func(regionID uint64, state *regionlock.LockedRangeState) {
-		targetTs := rt.staleLocksTargetTs.Load()
-		if state.ResolvedTs.Load() < targetTs && state.Initialized.Load() {
-			select {
-			case <-s.ctx.Done():
-			case s.resolveLockTaskCh <- resolveLockTask{
-				keyspaceID: span.KeyspaceID,
-				regionID:   regionID,
-				targetTs:   targetTs,
-				state:      state,
-				create:     time.Now(),
-			}:
-			// it is ok to ignore resolve lock task when the channel is full
-			default:
-				metrics.SubscriptionClientResolveLockTaskDropCounter.Inc()
-			}
-		}
-	}
-	return rt
-}
-
-func (s *subscriptionClient) GetResolvedTsLag() float64 {
-	pullerMinResolvedTs := uint64(0)
-	s.totalSpans.RLock()
-	for _, rt := range s.totalSpans.spanMap {
-		resolvedTs := rt.resolvedTs.Load()
-		if pullerMinResolvedTs == 0 || resolvedTs < pullerMinResolvedTs {
-			pullerMinResolvedTs = resolvedTs
-		}
-	}
-	s.totalSpans.RUnlock()
-	if pullerMinResolvedTs == 0 {
-		return 0
-	}
-	pdTime := s.pdClock.CurrentTime()
-	phyResolvedTs := oracle.ExtractPhysical(pullerMinResolvedTs)
-	lag := float64(oracle.GetPhysical(pdTime)-phyResolvedTs) / 1e3
-	return lag
-}
-
-func (r *subscribedSpan) resolveStaleLocks(targetTs uint64) {
-	util.MustCompareAndMonotonicIncrease(&r.staleLocksTargetTs, targetTs)
-	res := r.rangeLock.IterAll(r.tryResolveLock)
-	log.Debug("subscription client finds slow locked ranges",
-		zap.Uint64("subscriptionID", uint64(r.subID)),
-		zap.Any("ranges", res))
 }
